Added Map Test
Basically added the leaflet thing trying not to use the Google Maps API.
It will be a bit more of a pain in the ass but it is what it is. The Map
as of this commit shows a red pin emoji (📍) for the current location.
Requesting permissions from Android appears to work.
There is also a button that Zooms to current location. This is finicky
workaround to the BS that Google APIs wants me to do. No I am not going
to setup a billing profile and risk my card getting charged thousands of
dollars thank you very much. But yea its just leaflet with OSM map with
kind of a laggy feel.

All I need to do is add the Firestore DB integration. I'm thinking of
showing a random emoji (that can hopfully be changed when I add group
and user settings) for their location. There should be a little window
on the top left with the members and their coordinates. This is so much
but it'll keep me occupied.