Template for Adding Location and User panel
Basically I am going to store the locations in a `map` and its going to
be in the following format

```json
{
  "members": ["user1", "user2"],
  "locations": {
    "user1": {
      "geopoint": GeoPoint(37.7749, -122.4194),
      "timestamp": Timestamp
    },
    "user2": {
      "geopoint": GeoPoint(40.7128, -74.0060),
      "timestamp": Timestamp
    }
  }
}

```
Good night. I'm tired. According to Wakatime I've programmed for total
of 9 hours combined with yesterday and the few hours of today. I have
this weird soreness in my ribs hopefully that resolves. Or maybe because
I've been laying on my bed in an awkward position coding. But seriously,
good night Eshan.