Logs location to database
Finally logs to database but view doesn't update live visually. Need to
implement background functionality (tracking location in background)
but I'll do that last can't be that hard to do. Also need to add where
it stores a random emoji for each user (in firestore) and displays it in the scrollable
element. First things first though, I need to show pins for every user.
Right now, it only shows for the current user. Also `MapScreen.tsx` is
highly cluttered and I need to refactor a few things to clean it up and
make more readable so I can debug it.