Shows multiple pins for all users
Going to add clicking on the button of a given user, automatically zooms
to their location. Also the pins are the same red pin for all users. I
will implement it where a random emoji is assigned to a given user on
joining the group in the Firestore database. I will eventually add this
to the coming user settings menu.

Fixed issue in signup where it would store the actual value of what was
entered into the DB. Firestore automatically lowercases all usernames,
in the future I might allow casing but for now all usernames are
lowercase. Another thing to change is how fast the map loads up.