Added Admin Panel and Location Services Handling
The Admin Panel is where the emoji panel was, in the settings. It is now
a stack of the emoji settings and the admin settings. The admin can
remove any person in the group with the role of `member` as well as
promote members to `admin`. Admins cannot remove other admins from the
group. Admins can only see the admin panel option
and have access to the settings. Also added the ability for admins to
delete the group completly.

Also made sure that when signing up and logging in it trims the end and
start for whitespaces for both the username and password. Also made the
error alert for incorrect login.

Made it so if the location services are disabled it will show the last
known location of that user and show a warning at the top of the screen.

FEATURES TO ADD:
- Ability to leave group
- Better formatting on admin panel so it displays properly
- Add the ability for the user to change the location refresh time and
  increase the time for the default (updates to quickly right now)
- Add confirmation alert to remove members
- Add user settings
  - nickname (name displayed in group)
  - delete account (remove from groups)
- static -> dynamic code that can be initiated in the admin panel
  - assign unique group id for firestore document
  - create a collection of active group codes that have the join code
  - timestamp in the documents of expiration (30 minutes per code)